Finish project Spring boot

+++ src/main/java/com/dlea/Course/services/UserService.java
@@ -4,11 +4,16 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.dlea.Course.entities.User;
 import com.dlea.Course.repositories.UserRepository;
+import com.dlea.Course.services.exceptions.DatabaseException;
 import com.dlea.Course.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -23,7 +28,7 @@ public class UserService {
 	public User findById(Long id) {
 
 		Optional<User> obj = userRepository.findById(id);
-		return obj.orElseThrow(() -> new ResourceNotFoundException(id) );
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 
 	public User insert(User obj) {
@@ -31,13 +36,23 @@ public class UserService {
 	}
 
 	public User update(Long id, User obj) {
-		User entity = userRepository
-				.getReferenceById(id);/*
-										 * getReferenceById retorna uma referência à entidade identificada pelo ID fornecido, sem
-										 * realmente carregar seus dados do banco de dados até que seja necessário.
-										 */
-		updateDate(entity, obj);
-		return userRepository.save(entity);
+		
+		
+		try {
+
+			User entity = userRepository
+					.getReferenceById(id);/*
+											 * getReferenceById retorna uma referência à entidade identificada pelo ID
+											 * fornecido, sem realmente carregar seus dados do banco de dados até que
+											 * seja necessário.
+											 */
+			updateDate(entity, obj);
+			return userRepository.save(entity);
+
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id); 
+		}
+
 	}
 
 	private void updateDate(User entity, User obj) {
@@ -48,7 +63,17 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		userRepository.deleteById(id);
-	}
+		if (!userRepository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			userRepository.deleteById(id);
 
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+
+		}
+	}
 }
